ai_api/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.16
langsmith>=0.1.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.25
pydantic>=2.6.1
python-multipart>=0.0.9
